Rename AvailableGuardian to DecryptingGuardian Rename SubmittedBallot to EncryptedBallot, ciphertext_ballot.proto to encrypted_ballot.proto. Rename RecoveredPartialDecryption.guardianId to decryptingGuardianId update docs regenerate test files.
