+++ build.gradle.kts
@@ -172,7 +172,7 @@ val compileProtobuf =
             // TODO lame
             val commandLineStr =
                 "protoc --pbandk_out=./src/commonMain/kotlin/ --proto_path=./src/commonMain/proto " +
-                    "ciphertext_ballot.proto ciphertext_tally.proto common.proto " +
+                    "encrypted_ballot.proto ciphertext_tally.proto common.proto " +
                     "election_record.proto election_record_ver1.proto manifest.proto " +
                     "plaintext_ballot.proto plaintext_tally.proto " +
                     "trustees.proto"
