+++ docs/commandLineInterface.md
@@ -1,0 +1,8 @@
+# Command Line Programs
+
+draft 5/17/2022 for proto_version = 2.0.0
+
+## Generate electionConfig
+
+## Run Batch Encryption
+
