+++ src/commonMain/proto/election_record.proto
@@ -62,11 +62,11 @@ message TallyResult {
 message DecryptionResult {
   TallyResult tally_result = 1;
   PlaintextTally decrypted_tally = 2;
-  repeated AvailableGuardian decrypting_guardians = 3;
+  repeated DecryptingGuardian decrypting_guardians = 3;
   map<string, string> metadata = 4;
 }
 
-message AvailableGuardian {
+message DecryptingGuardian {
   string guardian_id = 1;
   uint32 x_coordinate = 2; // x coordinate in the secret polynomial. must be > 0
   ElementModQ lagrange_coefficient = 3;
